refactor: extract matcher and github actions crates
Move the shared wildcard matcher into a dependency-free crate and give GitHub Actions
workflow handling its own library. Keep command opt-in policy in the CLI and preserve
pnpm_config::matcher through a reexport.

Remove the configuration dependency from workspace filtering and move YAML dependencies
from the CLI to the GitHub Actions crate. Preserve behavior and existing tests.

Follow-up to pnpm/pnpm#14804.
