+++ pnpm/crates/default-reporter/Cargo.toml
@@ -11,6 +11,7 @@ license.workspace     = true
 repository.workspace  = true
 
 [dependencies]
+pnpm-matcher  = { workspace = true }
 pnpm-config   = { workspace = true }
 pnpm-reporter = { workspace = true }
