+++ pnpm/crates/deps-inspection-peers/Cargo.toml
@@ -11,6 +11,7 @@ license.workspace     = true
 repository.workspace  = true
 
 [dependencies]
+pnpm-matcher                           = { workspace = true }
 pnpm-catalogs-resolver                 = { workspace = true }
 pnpm-catalogs-types                    = { workspace = true }
 pnpm-config                            = { workspace = true }
