+++ pnpm/crates/deps-restorer/Cargo.toml
@@ -11,6 +11,7 @@ license.workspace     = true
 repository.workspace  = true
 
 [dependencies]
+pnpm-matcher                  = { workspace = true }
 rustc-hash                    = { workspace = true }
 pnpm-cmd-shim                 = { workspace = true }
 pnpm-config                   = { workspace = true }
