+++ pnpm/crates/matcher/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name                 = "pnpm-matcher"
+version              = "0.0.1"
+publish              = false
+authors.workspace    = true
+description          = "Ordered wildcard matching with include and exclude patterns"
+edition.workspace    = true
+homepage.workspace   = true
+keywords.workspace   = true
+license.workspace    = true
+repository.workspace = true
+
+[lints]
+workspace = true
