+++ pnpm/crates/package-manager/Cargo.toml
@@ -11,6 +11,7 @@ license.workspace     = true
 repository.workspace  = true
 
 [dependencies]
+pnpm-matcher                           = { workspace = true }
 rustc-hash                             = { workspace = true }
 pnpm-catalogs-config                   = { workspace = true }
 pnpm-catalogs-protocol-parser          = { workspace = true }
